+++ src/services/addOn/awsGathering.service.ts
@@ -165,7 +165,6 @@ function addRegion(resources:any, region:string) {
 }
 
 async function ec2SGListing(client: EC2, region: string): Promise<any> {
-    if(!currentConfig.ObjectNameNeed?.includes("ec2SG")) return null;
     try {
         const data = await client.describeSecurityGroups().promise();
         let jsonData = JSON.parse(JSON.stringify(data.SecurityGroups));
@@ -179,7 +178,6 @@ async function ec2SGListing(client: EC2, region: string): Promise<any> {
 }
 
 async function ec2VolumesListing(client: EC2, region: string): Promise<any> {
-    if(!currentConfig.ObjectNameNeed?.includes("ec2Volume")) return null;
     try {
         const data = await client.describeVolumes().promise();
         let jsonData = JSON.parse(JSON.stringify(data.Volumes));
@@ -193,7 +191,6 @@ async function ec2VolumesListing(client: EC2, region: string): Promise<any> {
 }
 
 async function ec2InstancesListing(client: EC2, region: string): Promise<Array<EC2.Instance> | null> {
-    if(!currentConfig.ObjectNameNeed?.includes("ec2Instance")) return null;
     try {
         const data = await client.describeInstances().promise();
         let jsonData = JSON.parse(JSON.stringify(data.Reservations));
@@ -207,7 +204,6 @@ async function ec2InstancesListing(client: EC2, region: string): Promise<Array<E
 }
 
 async function rdsInstancesListing(client: RDS, region: string): Promise<any> {
-    if(!currentConfig.ObjectNameNeed?.includes("rds")) return null;
     try {
         const data = await client.describeDBInstances().promise();
         let jsonData = JSON.parse(JSON.stringify(data.DBInstances));
@@ -221,7 +217,6 @@ async function rdsInstancesListing(client: RDS, region: string): Promise<any> {
 }
 
 async function resourceGroupsListing(client: ResourceGroups, region: string): Promise<any> {
-    if(!currentConfig.ObjectNameNeed?.includes("resourceGroup")) return null;
     try {
         const data = await client.listGroups().promise();
         let jsonData = JSON.parse(JSON.stringify(data.Groups));
@@ -235,7 +230,6 @@ async function resourceGroupsListing(client: ResourceGroups, region: string): Pr
 }
 
 async function tagsValueListing(client: ResourceGroupsTaggingAPI, region: string): Promise<any> {
-    if(!currentConfig.ObjectNameNeed?.includes("tagsValue")) return null;
     try {
         interface TagParams {Key: string;}
         const dataKeys = await client.getTagKeys().promise();
@@ -257,7 +251,6 @@ async function tagsValueListing(client: ResourceGroupsTaggingAPI, region: string
 }
 
 async function s3BucketsListing(client: S3, region: string): Promise<Array<S3> | null> {
-    if(!currentConfig.ObjectNameNeed?.includes("s3")) return null;
     try {
         const data = await client.listBuckets().promise();
         let jsonData = JSON.parse(JSON.stringify(data.Buckets));
@@ -271,7 +264,6 @@ async function s3BucketsListing(client: S3, region: string): Promise<Array<S3> |
 }
 
 async function ecsClusterListing(client: ECS, region: string): Promise<any> {
-    if(!currentConfig.ObjectNameNeed?.includes("ecsCluster")) return null;
     try {
         const data = await client.describeClusters().promise();
         let jsonData = JSON.parse(JSON.stringify(data.clusters));
@@ -285,7 +277,6 @@ async function ecsClusterListing(client: ECS, region: string): Promise<any> {
 }
 
 async function ecrImagesListing(client: ECR, region: string): Promise<any> {
-    if(!currentConfig.ObjectNameNeed?.includes("ecrImage")) return null;
     try {
         const data = await client.describeRepositories().promise();
         let jsonData = JSON.parse(JSON.stringify(data.repositories));